feat: vim.ui.select() interface for :WorkspacesOpen
When :WorkspacesOpen is run without an argument, vim.ui.select() will be
run prompting for a workspace to select.

Because the functions in vim.ui are meant to be user-defined, this
allows workspace picking with a user's preferred list picker instead of
requiring Telescope for a nice UI. Or perhaps the vim.ui.select()
default is also preferred!